requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
fudge==1.1.0
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
pkg-resources==0.0.0
PyYAML==3.12
requests==2.18.1
simplejson==3.11.1
soundcloud==0.5.0
urllib3==1.21.1
Werkzeug==0.12.2
